mv core if it exists only

+++ run-valgrind.sh
@@ -5,7 +5,9 @@ ulimit -c unlimited
 while true; do
 echo 'Starting ./frogserv'
 valgrind -v --leak-check=full --show-reachable=yes ./frogserv 2>&1 | tee frogserv-valgrind-`date +"%F_%H-%M-%S"`.log
-mv core core-`date +"%F_%H-%M-%S"`
+if [ -f core ]; then
+	mv core core-`date +"%F_%H-%M-%S"`
+fi
 if [ ! -z $1 ]; then
 /usr/sbin/sendmail -t "$1" << EOF
 From: frogserv@mappinghell.net
